Revised aero breakpoints to remove possibly incorrect values
beta -7 instead of -10
removed alpha 26 and 28 - so it now has 24, 30 - which will help with some of the larger spikes on yaw. I'm not sure if these are right or not, but it looks like it could be wrong. I experimented with model changes and a few runs in OpenVSP and it still isn't clear if these spikes are related to poor mesh resolution or are genuine. In any case at this alpha, with the way that OpenVSP works (i.e. no turbulent flow), it is more likely that the spikes are numerical inaccuracy than real effects - so until more information comes to light this seems to be a reasonable compromise.