Revised aero model.
Changed to have a root airflow (visually matched) with an LE sweep of 40 degrees that is similar to the aircraft (we can't model this accurately within VSP).

Adjusted CG (previously too far forwards)

Better matches the one Cm diagram we have.

Without FBW this is twitchy but controllable, which gives me hope because reports indicate that the aircraft could be flyable with reduced FBW.